+++ src/CMakeLists.txt
@@ -4,12 +4,12 @@ add_executable(CIMU
   file_reader.c
   file_writer.c
   angles.c
+  filters/iir.c
 )
 
 target_link_libraries(CIMU PRIVATE CMSISDSP)
 # # math standart
 target_link_libraries(CIMU PRIVATE m)
-
 
 #
 target_include_directories (CIMU PRIVATE
